# Synchronization updates: final and volatile fields, immutable struct