--- a/lib/concurrent/synchronization.rb
+++ b/lib/concurrent/synchronization.rb
@@ -7,22 +7,21 @@ require 'concurrent/synchronization/rbx_object'
 
 module Concurrent
   module Synchronization
-    class Object < case
-                           when Concurrent.on_jruby?
-                             JavaObject
+    Implementation = case
+                     when Concurrent.on_jruby?
+                       JavaObject
+                     when Concurrent.on_cruby? && (RUBY_VERSION.split('.').map(&:to_i) <=> [1, 9, 3]) <= 0
+                       MonitorObject
+                     when Concurrent.on_cruby?
+                       MutexObject
+                     when Concurrent.on_rbx?
+                       RbxObject
+                     else
+                       MutexObject
+                     end
+    private_constant :Implementation
 
-                           when Concurrent.on_cruby? && (RUBY_VERSION.split('.').map(&:to_i) <=> [1, 9, 3]) <= 0
-                             MonitorObject
-
-                           when Concurrent.on_cruby?
-                             MutexObject
-
-                           when Concurrent.on_rbx?
-                             RbxObject
-
-                           else
-                             MutexObject
-                           end
+    class Object < Implementation
     end
   end
 end