--- a/lib/concurrent/synchronization.rb
+++ b/lib/concurrent/synchronization.rb
@@ -27,3 +27,5 @@ module Concurrent
     end
   end
 end
+
+require 'concurrent/synchronization/immutable_struct'