- Add `ns_initialize` for ivars setting  inside synchronize block
- cleans up the Synchronization::Object space
- add `ensure_ivar_visibility!` for final fields
- add volatile attributes