@jdantonio 2015-04-30
:+1: I'm :ok_hand: with merging this. I'll rebase and continue working on #271, #283, and #280 afterwards.