@pitr-ch 2015-05-02
The failing test is not using `Synchronized::Object` and I cannot get it reproduced :/ If It succeeds at least once on travis I'll merge this branch and add it to intermittently failing tests.