@jdantonio 2015-05-02
I would prefer to merge this branch then work on the failing test afterwards:

1. This code needs to be in master for is to proceed with our other work
1. As noted, the failing test appears to be unrelated to the changes in this PR
1. I spent over an hour the other day trying to get Rbx to install on an Ubuntu laptop and wasn't successful, so fixing this test may be very difficult