@pitr-ch 2015-05-02
Merging since travis gave a different intermittent error.